python/requirements.txt
GitPython==3.1.43
h5py==3.4.0
k_means_constrained==0.7.3
matplotlib==3.4.3
netgraph==4.13.2
numpy==1.24.4
pandas==2.0.3
PyYAML==6.0.2
scikit_learn==0.24.2
scipy==1.14.1
Shapely==2.0.6
umap_learn==0.5.3